BOT TELEGRAM V0 FINISHED

+++ backend/apps/account/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class User(AbstractUser):
     username = models.CharField(max_length=30, unique=True)
     email = models.EmailField(unique=True, blank=True, null=True)
-    telegram_id = models.IntegerField(null=True, unique=True)
+    telegram_id = models.BigIntegerField(null=True, unique=True)
     first_name = models.CharField(max_length=30, blank=True, null=True)
     last_name = models.CharField(max_length=30, blank=True, null=True)
     mobile = models.CharField(max_length=30, blank=True, null=True)
